banking_loan_approval2/requirements.txt
numpy==1.18.4
scikit-learn==0.23.1
pandas==0.25.3
cortex-certifai-model-sdk
cortex-certifai-common
